requirements.txt
beautifulsoup4==4.12.3
contractions==0.1.73
groq==0.11.0
nltk==3.8.1
nltk==3.9.1
python-dotenv==1.0.1
scikit_learn==1.2.0
scikit_learn==1.3.0
selenium==4.25.0
spacy==3.8.2
streamlit==1.39.0
